pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "loaf-control"
version = "0.1.0"
description = "Sourdough baking log and analysis"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
]

[tool.setuptools]
packages = []

[dependency-groups]
dev = [
    "ruff>=0.11.7",
    "pytest>=8.0.0",
    "pre-commit>=4.0.0",
]

[tool.ruff]
line-length = 119
indent-width = 4
exclude = [
    ".venv",
    "build",
    "dist",
    ".git",
    "__pycache__"
]

[tool.ruff.lint]
select = ["I", "E", "F", "W"]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
